pairings CRUD

+++ RATAPPLibrary/Services/BreedingService.cs
@@ -30,6 +30,7 @@ namespace RATAPPLibrary.Services
 
         }
 
+        //PAIRINGS
         //get all pairings
         public async Task<List<Pairing>> GetAllPairingsAsync()
         {
@@ -55,11 +56,71 @@ namespace RATAPPLibrary.Services
         }
 
         //get all upcoming pairings (no pairing start, or end date)
-        //get all past pairings (pairing start and end date) 
+        public async Task<List<Pairing>> GetAllUpcomingPairingsAsync()
+        {
+            // Retrieve active pairings where PairingEndDate is null
+            var futurePairings = await _context.Pairing
+                                               .Where(p => p.PairingEndDate == null && p.PairingStartDate == null)
+                                               .ToListAsync();
+
+            // Return the list, which may be empty if no future pairings exist
+            return futurePairings;
+        }
+
+        //get all past pairings (pairing start and end date)
+        public async Task<List<Pairing>> GetAllPastPairingsAsync()
+        {
+            // Retrieve active pairings where PairingEndDate is null
+            var pastPairings = await _context.Pairing
+                                               .Where(p => p.PairingEndDate != null && p.PairingStartDate != null)
+                                               .ToListAsync();
+
+            // Return the list, which may be empty if no past pairings exist
+            return pastPairings;
+        }
+
+        //get all pairings for animal id
+        //looking for dam, or sire with matching ID
+        public async Task<List<Pairing>> GetAllPairingsByAnimalIdAsync(int animalID)
+        {
+            // Retrieve active pairings where PairingEndDate is null
+            var pastPairings = await _context.Pairing
+                                               .Where(p => p.SireId == animalID || p.DamId == animalID)
+                                               .ToListAsync();
+
+            // Return the list, which may be empty if id doesn't exist
+            return pastPairings;
+        }
 
         //get all pairings for line
-        //get all pairings for group
-        //get all pairings for animal id
+        public async Task<List<Pairing>> GetAllPairingsByLineIdAsync(int lineID)
+        {
+            //probably should get pairing, check for project then get line id but TODO this works for now 
+
+            // Retrieve active pairings where PairingEndDate is null
+            var linePairings = await _context.Pairing
+                                               .Where(p => p.Project.LineId == lineID)
+                                               .ToListAsync();
+
+            // Return the list, which may be empty if lineID doesn't exist
+            return linePairings;
+        }
+
+        //get pairings by species 
+        public async Task<List<Pairing>> GetAllPairingsBySpeciesAsync(string species)
+        {
+            //probably should get pairing, check for project then get line, stock, species first to check if its populated but TODO this works for now 
+
+            // Retrieve active pairings where PairingEndDate is null
+            var pastPairings = await _context.Pairing
+                                               .Where(p => p.Project.Line.Stock.Species.CommonName == species)
+                                               .ToListAsync();
+
+            // Return the list, which may be empty if id doesn't exist
+            return pastPairings;
+        }
+
+        //get all pairings for group TODO
 
         //add new pairing 
         //individual variables TODO may remove this and switch to just using objects 
@@ -116,9 +177,6 @@ namespace RATAPPLibrary.Services
                 DateTime createdOn = DateTime.Now;
                 DateTime lastUpdated = DateTime.Now;
 
-                //map to a pairing object No need to map if we pass in an object 
-                //Pairing pairing = mapToPairingObject(pairingId, damId, sireId, projectId, createdOn, lastUpdated, startDate, endDate);
-
                 //make a new entry in the animal pairing table for the specified dam and sire 
                 _context.Pairing.Add(pair);
             }
@@ -155,14 +213,25 @@ namespace RATAPPLibrary.Services
 
 
         //delete pairing 
-        //update pairing 
-        //find pairing 
+        //TODO 
+
+        //update pairing
+        //TODO 
+        //cannot update pairing id as this would conflict with litters 
+        //you can only update certain pieces of a pairing such as start and end date once a litter has been associated with a pairing
+        //so, this must be checked for i.e. is there a litter associated with this pairing? 
+        /// <summary>
+       
+        /// </summary>
+        /// <returns></returns>
+        //
 
 
-        //litters 
+
+        //LITTERS
 
         //when a litter is added there should be pups associated with it which implies that new animals will be created 
-       
+
         //get all litters 
         public async Task<List<Litter>> GetAllLittersAsync()
         {
@@ -204,6 +273,9 @@ namespace RATAPPLibrary.Services
         }
 
         //update litter, individual variable not object 
+        //TODO - litter ID cannot be updated as this would conflict with associated pups
+        //this is the basic update method I need a method for updating animals associated with litters
+        //i.e. adding or deleting animals from litters 
         public async Task<Litter> UpdateLitterAsync(int litterId, string name, DateTime dob, int numPups)
         {
             var litter = await _context.Litter.FindAsync(litterId);
@@ -219,17 +291,29 @@ namespace RATAPPLibrary.Services
             return litter;
         }
 
+        //TODO
+        //update pups in litter (actual animal objects)
+
+        //add 
+        //delete
+        //an animal doesn't have to be associated with a litter, so the user should be able to "remove animal" from a litter
+        //the user should also be able to "add animal" to a litter
+
+
         //delete litter
+        //TODO: for now, only allow this to happen if there are no animals associated with the litter 
         public async Task DeleteLitterAsync(int litterId)
         {
             var litter = await _context.Litter.FindAsync(litterId);
 
             if (litter == null) throw new KeyNotFoundException($"Litter {litterId} not found");
+            if (litter.Animals != null) throw new InvalidOperationException($"Litter {litterId} has pups associated with it so it cannot be deleted"); //TODO need to set this up, eventually there should just be a warning about deleting litters with pups associated 
+            
             _context.Litter.Remove(litter);
 
             await _context.SaveChangesAsync();
         }
 
-        //TODO when a litter is added, new animals should be created for said pups ?
+        //TODO when a litter is added, new animals should be created for said pups ? - future feature 
     }
 }
